+++ src/MonacoContainer/index.spec.js
@@ -1,12 +1,17 @@
 import React from 'react';
-import DiffEditor from '.';
+import MonacoContainer from '.';
 import { render } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
 
-describe('<DiffEditor />', () => {
+describe('<MonacoContainer />', () => {
   it('should check render with snapshot', () => {
     const component = render(
-      <DiffEditor />,
+      <MonacoContainer
+      	width="100%"
+      	height="100vh"
+      	loading="loading..."
+      	isEditorReady={false}
+      />,
     );
 
     expect(component).toMatchSnapshot();